keep virtualenvs in home/.virtualenv
Bash has a problem with spaces in the shebang (the #! at the top of a
file). So putting the virtualenv in the jobs workspace meant that the
pip executables shebang could be in a path which has spaces.

To fix this I've changed the jenkins script to create virtualenvs in the
home/.virtualenv folder and each virtualenv is named the job name with
spaces replaced with underscores.